requirements.txt
Flask
pandas
psycopg2-binary
openai
celery
redis
